Moved tutorial components from onboarding module.
Moves the components and constants from overview/onboarding to
options/tutorial.
