+++ src/options/tutorial/components/FeatureInfo.css
